+++ src/ldp.rs
@@ -185,7 +185,7 @@ impl LinkedDataProofs {
             } => match &curve[..] {
                 "Ed25519" => {
                     if let Some(ref vm) = options.verification_method {
-                        if vm.starts_with("did:tz") {
+                        if vm.starts_with("did:tz:") || vm.starts_with("did:pkh:tz:") {
                             return Ed25519BLAKE2BDigestSize20Base58CheckEncodedSignature2021
                                 .sign(document, options, &key)
                                 .await;
@@ -234,7 +234,7 @@ impl LinkedDataProofs {
                     }
                     "P-256" => {
                         if let Some(ref vm) = options.verification_method {
-                            if vm.starts_with("did:tz") {
+                            if vm.starts_with("did:tz:") || vm.starts_with("did:pkh:") {
                                 return P256BLAKE2BDigestSize20Base58CheckEncodedSignature2021
                                     .sign(document, options, &key)
                                     .await;
@@ -269,7 +269,7 @@ impl LinkedDataProofs {
             }
             Algorithm::EdDSA => {
                 if let Some(ref vm) = options.verification_method {
-                    if vm.starts_with("did:tz") {
+                    if vm.starts_with("did:tz:") || vm.starts_with("did:pkh:tz:") {
                         return Ed25519BLAKE2BDigestSize20Base58CheckEncodedSignature2021
                             .prepare(document, options, public_key)
                             .await;
@@ -286,7 +286,7 @@ impl LinkedDataProofs {
             }
             Algorithm::ES256 => {
                 if let Some(ref vm) = options.verification_method {
-                    if vm.starts_with("did:tz") {
+                    if vm.starts_with("did:tz:") || vm.starts_with("did:pkh:tz:") {
                         return P256BLAKE2BDigestSize20Base58CheckEncodedSignature2021
                             .prepare(document, options, public_key)
                             .await;